Implementiert Funktion zur Anzeige eines einzelnen Ereignisses nach ID

+++ EventManagmentSystem/Models/ViewModel/EventViewModel.cs
@@ -10,7 +10,7 @@ namespace EventManagmentSystem.Models.ViewModel
 
         public DateTime Date { get; set; }
 
-        public double Price { get; set; } //keine Unterscheidung für Plätze 
+        public decimal Price { get; set; } //keine Unterscheidung für Plätze 
 
         public EventStateEnum State { get; set; }
